feat!(twext): use iterators instead of slices
The main us case is to process a slice of entries by filtering and
aggregating the entries. At each stage the whole slice is passed. For
huge slices, this wastes memory.

This commit improves memory usage by changing the main structure to be
iterators. The filter and aggregation helpers are changed to receive and
return iterators.

Still todo is to change the input JSON decoding into a streaming
operation.